テンプレートの動的変換機能を追加し、AccessorCreationServiceを更新
