+++ src/main/java/kmg/tool/application/service/DynamicTemplateConversionService.java
@@ -1,0 +1,12 @@
+package kmg.tool.application.service;
+
+import kmg.tool.domain.service.Two2OneService;
+
+/**
+ * テンプレートの動的変換サービスインタフェース
+ */
+public interface DynamicTemplateConversionService extends Two2OneService {
+
+    // 処理なし
+
+}
