+++ src/main/java/kmg/tool/application/service/impl/AccessorCreationServiceImpl.java
@@ -1,102 +1,18 @@
 package kmg.tool.application.service.impl;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-
 import org.springframework.stereotype.Service;
 
-import kmg.core.infrastructure.type.KmgString;
-import kmg.core.infrastructure.types.KmgDelimiterTypes;
 import kmg.tool.application.service.AccessorCreationService;
-import kmg.tool.domain.types.KmgToolGenMessageTypes;
 import kmg.tool.infrastructure.exception.KmgToolException;
 
 /**
- * アクセサ作成サービス実装クラス<br>
+ * アクセサ作成サービス<br>
  *
  * @author KenichiroArai
  */
 @Service
-public class AccessorCreationServiceImpl implements AccessorCreationService {
-
-    /** 入力ファイルパス */
-    private Path inputPath;
-
-    /** テンプレートファイルパス */
-    private Path templatePath;
-
-    /** 出力ファイルパス */
-    private Path outputPath;
-
-    /**
-     * 入力ファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return 入力ファイルパス
-     */
-    @Override
-    public Path getInputPath() {
-
-        final Path result = this.inputPath;
-        return result;
-
-    }
-
-    /**
-     * 出力ファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return 出力ファイルパス
-     */
-    @Override
-    public Path getOutputPath() {
-
-        final Path result = this.outputPath;
-        return result;
-
-    }
-
-    /**
-     * 初期化する
-     *
-     * @return true：成功、false：失敗
-     *
-     * @param inputPath
-     *                     入力ファイルパス
-     * @param templatePath
-     *                     テンプレートファイルパス
-     * @param outputPath
-     *                     出力ファイルパス
-     */
-    @SuppressWarnings("hiding")
-    @Override
-    public boolean initialize(final Path inputPath, final Path templatePath, final Path outputPath) {
-
-        final boolean result = true;
-
-        this.inputPath = inputPath;
-        this.templatePath = templatePath;
-        this.outputPath = outputPath;
-
-        return result;
-
-    }
+public class AccessorCreationServiceImpl extends DynamicTemplateConversionServiceImpl
+    implements AccessorCreationService {
 
     /**
      * 処理する
@@ -111,73 +27,7 @@ public class AccessorCreationServiceImpl implements AccessorCreationService {
 
         boolean result = false;
 
-        /* テンプレートの取得 */
-        String template = null;
-
-        try {
-
-            template = Files.readAllLines(this.templatePath).stream()
-                .collect(Collectors.joining(KmgDelimiterTypes.LINE_SEPARATOR.get()));
-
-        } catch (final IOException e) {
-
-            // 例外をスローする
-            // TODO KenichiroArai 2025/03/06 ログメッセージ
-            final KmgToolGenMessageTypes msgType     = KmgToolGenMessageTypes.NONE;
-            final Object[]               messageArgs = {
-                this.templatePath.toString()
-            };
-            throw new KmgToolException(msgType, messageArgs, e);
-
-        }
-
-        /* 入力から出力の処理 */
-        try (final BufferedReader brInput = Files.newBufferedReader(this.inputPath);
-            final BufferedWriter bw = Files.newBufferedWriter(this.outputPath);) {
-
-            String output = template;
-            String line   = null;
-
-            while ((line = brInput.readLine()) != null) {
-
-                line = line.replace("final", KmgString.EMPTY);
-                line = line.replace("static", KmgString.EMPTY);
-                final Pattern patternComment = Pattern.compile("/\\*\\* (\\S+)");
-                final Matcher matcherComment = patternComment.matcher(line);
-
-                if (matcherComment.find()) {
-
-                    output = output.replace("$name", matcherComment.group(1));
-                    continue;
-
-                }
-
-                final Pattern patternSrc = Pattern.compile("private\\s+((\\w|\\[\\]|<|>)+)\\s+(\\w+);");
-                final Matcher matcherSrc = patternSrc.matcher(line);
-
-                if (matcherSrc.find()) {
-
-                    output = output.replace("$type", matcherSrc.group(1));
-                    output = output.replace("$item", matcherSrc.group(3));
-                    output = output.replace("$Item",
-                        matcherSrc.group(3).substring(0, 1).toUpperCase() + matcherSrc.group(3).substring(1));
-                    bw.write(output);
-                    bw.write(System.lineSeparator());
-                    output = template;
-
-                }
-
-            }
-
-            result = true;
-
-        } catch (final IOException e) {
-
-            // TODO KenichiroArai 2025/03/06 ログメッセージ
-            final KmgToolGenMessageTypes msgType = KmgToolGenMessageTypes.NONE;
-            throw new KmgToolException(msgType, e);
-
-        }
+        result = super.process();
 
         return result;
