--- src/main/resources/tool/io/template/AccessorCreationlTool.txt
@@ -1,22 +1,0 @@
-/**
- * $nameを設定する<br>
-  * @author KenichiroArai
-  * @sine 1.0.0
-  * @version 1.0.0
- * @param $item $name
- */
- public void set$Item($type $item) {
-    this.$item = $item;
- }
-
- /**
-  * $nameを返す<br>
-  * @author KenichiroArai
-  * @sine 1.0.0
-  * @version 1.0.0
-  * @return $name
-  */
- public $type get$Item() {
-    $type result = this.$item;
-    return result;
- }
