+++ src/main/resources/tool/io/template/AccessorCreationlTool.yml
@@ -1,0 +1,33 @@
+columns:
+  - key: "名称"
+    placeholder: "{name}"
+  - key: "項目"
+    placeholder: "{item}"
+  - key: "項目先頭大文字"
+    placeholder: "{Item}"
+  - key: "型"
+    placeholder: "{type}"
+
+template: |
+    /**
+     * {name}を設定する<br>
+      * @author KenichiroArai
+      * @sine 0.1.0
+      * @version 0.1.0
+     * @param {item} {name}
+     */
+     public void set{Item}({type} {item}) {
+        this.{item} = {item};
+     }
+    
+     /**
+      * {name}を返す<br>
+      * @author KenichiroArai
+      * @sine 0.1.0
+      * @version 0.1.0
+      * @return {name}
+      */
+     public {type} get{Item}() {
+        {type} result = this.{item};
+        return result;
+     }
